tutoriaalin implementaatio
