--- Assets/scripts/musicControlTutorial.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6d7191373ec68af49a5151bfd7172b47
