+++ Assets/scripts/tutorial/loadArea.cs
@@ -6,6 +6,7 @@ public class loadArea : MonoBehaviour
     public new Collider collider;
     public string prefix;
     public instructionHandler instructionHandler; //jotta tää on paljon helpompaa
+    public musicControlTutorial musicControlTutorial;
 
 
 
@@ -14,7 +15,7 @@ public class loadArea : MonoBehaviour
         prefix = gameObject.name.Substring(0, 2);
         collider = GetComponent<Collider>();
         instructionHandler = GameObject.Find("instructionHandler").GetComponent<instructionHandler>();
-
+        musicControlTutorial = GameObject.Find("music").GetComponent<musicControlTutorial>();
     }
 
     private void OnTriggerEnter(Collider other)
@@ -48,6 +49,22 @@ public class loadArea : MonoBehaviour
                 , 1);
                 
                 //StartCoroutine(ChangeAnimOverrides("driving:1", 2));
+                StartCoroutine(FadeDeath(1.0f));
+                break;
+            case "50":
+                musicControlTutorial.MusicSections("2_FINAL_TUTORIAL_main");
+                StartCoroutine(FadeDeath(1.0f));
+                break;
+            case "51":
+                musicControlTutorial.MusicSections("3_FINAL_TUTORIAL_main");
+                StartCoroutine(FadeDeath(1.0f));
+                break;
+            case "52":
+                musicControlTutorial.MusicSections("4_FINAL_TUTORIAL_main");
+                StartCoroutine(FadeDeath(1.0f));
+                break;
+            case "53":
+                musicControlTutorial.MusicSections("5_FINAL_TUTORIAL_main");
                 StartCoroutine(FadeDeath(1.0f));
                 break;
             default:
